backend/requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-google-genai
sentence-transformers
faiss-cpu
